% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateMeanDays.R
\name{interpolateMeanDays}
\alias{interpolateMeanDays}
\title{Interpolate mean days over the coordinates extent}
\usage{
interpolateMeanDays(Lon, Lat, meandays)
}
\arguments{
\item{Lon}{a vector of Longitude values}

\item{Lat}{a vector of Latitude values}

\item{meandays}{a vector of values to be interpolatd}
}
\value{
A raster of interpolaed mean days values covering the extent of 
    the coordinates
}
\description{
This function interpolates the mean days values over the extent of the input 
coordinates.
The function raises an error if the length of the inputs differ
}
