% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc4fullpaths.R
\name{nc4fullpaths}
\alias{nc4fullpaths}
\title{Create vector of fullpaths to netCDF files in the folder}
\usage{
nc4fullpaths(folder)
}
\arguments{
\item{folder}{Path to folder with netCDF files}
}
\value{
A dataframe of extracted temperature values
}
\description{
This function get the full path to all netCDF files in a folder. The 
function is recursive: it will search the sub folders.
}
