% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewTable.R
\name{viewTable}
\alias{viewTable}
\title{View data per Variety}
\usage{
viewTable(inputData, table.title = "'Variety Table'", variety = 1)
}
\arguments{
\item{inputData}{a data frame, or filepath to a xlsx file, generated 
by MSPP tool or agMERRAnc4ToCSV::mergeSummarySheets().}
}
\value{
A plot of the data in viewer that can be exported or saved.
}
\description{
This function views MSPP output as a table of variety with each "Name 2" 
location as a column heading. The rows contain data per date
}
