% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsSpatialPointsDataFrame.R
\name{wgsSpatialPointsDataFrame}
\alias{wgsSpatialPointsDataFrame}
\title{Create a SpatialPointsDataframe in WGS coordinate system fron a CSV file}
\usage{
wgsSpatialPointsDataFrame(filepath, crsobject, LongitudeColumn, LatitudeColumn)
}
\arguments{
\item{filepath}{path to csv file with coordianes}

\item{crsobject}{object of 
class \CRS{http://127.0.0.1:35826/help/library/sp/help/CRS}}
}
\value{
object of class SpatialPointsDataframe in WGS coordinate system
}
\description{
This function returns a SpatialPointsDataframe in WGS coordinate system 
generated from a csv file with coordinates in any system. This function 
removes entries in the csv with NAs
}
