% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtrctagMERRAnc4.R
\name{xtrctagMERRAnc4}
\alias{xtrctagMERRAnc4}
\title{Extract a netCDF file downloaded from agMERRA}
\usage{
xtrctagMERRAnc4(nc4Files, shpdf, name_1, name_2)
}
\arguments{
\item{nc4Files}{a vector of fullpaths to netCDF files}

\item{shpdf}{Object of class SpatialPointsDataFrame}

\item{name_1}{column name in shpdf to be added in the output}

\item{name_2}{column name in shpdf to be added in the output}
}
\value{
A dataframe of extracted temperature values
}
\description{
This function extracts data from a netCDF file using a 
SpatialPointsDataframe and ouptus the data in a format usable in MSPP tool
}
